requirements.txt
anyio==4.4.0
Arpeggio==2.0.2
attrs==23.2.0
black==24.4.2
certifi==2024.6.2
cfgv==3.4.0
click==8.1.7
dill==0.3.8
distlib==0.3.8
filelock==3.14.0
grpcio==1.60.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
identify==2.5.36
idna==3.7
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==1.26.4
packaging==24.0
parver==0.5
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
pre-commit==3.7.1
protobuf==4.25.3
pulumi==3.119.0
pulumi_gcp==7.26.0
python-dotenv==1.0.1
python-telegram-bot==21.3
PyYAML==6.0.1
semver==2.13.0
six==1.16.0
sniffio==1.3.1
virtualenv==20.26.2
